+++ RELEASES.md
@@ -1,6 +1,9 @@
 See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 
 `2.0.7-SNAPSHOT`
+* prov-jsonld: fixed context url
+* prov-model: builder for bundles
+* prov-template-compiler: generating outstanding sql methods and tables
 
 `2.0.6`
 * prov-template-compiler: introduced sqlify for a few more keywords